global_helpers should contain only global functions
global functions are functions that are called from multiple unrelated files
