+++ src/pages/create-profile/helpers.ts
@@ -1,0 +1,12 @@
+import randomColor from "randomcolor";
+
+export function getRandomColor() {
+	return randomColor({
+		luminosity: "random",
+		hue: "random",
+	});
+}
+
+export function getRandomInt(max: number) {
+	return Math.floor(Math.random() * max);
+}
